# solution/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления данными клиентов, рекламодателей, рекламными кампаниями, показом объявлений, статистикой и управлением \"текущим днём\" в системе.",
        "title": "PROD Backend 2025 Advertising Platform API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/advertisers/campaigns/generate": {
            "post": {
                "description": "Генерирует текст рекламы на основе имени рекламодателя и названии рекламы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Генерация текста рекламы",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GenerateAdTextResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/advertisers/campaigns/moderation": {
            "patch": {
                "description": "Переключает статус модерации. Может быть true или false",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Переключение модерации",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwitchModerationResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/advertisers/{advertiserId}/campaigns": {
            "get": {
                "description": "Возвращает кампании рекламодателя по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Получение кампаний рекламодателя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID рекламодателя",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Campaign"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает рекламную кампанию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Создание кампании",
                "parameters": [
                    {
                        "description": "CampaignRequest",
                        "name": "CreateCampaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CampaignRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Advertiser ID",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/advertisers/{advertiserId}/campaigns/{campaignId}": {
            "get": {
                "description": "Возвращает кампанию по ее ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Получение кампании",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет разрешённые параметры рекламной кампании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Обновление кампании",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет рекламную кампанию по ее ID",
                "tags": [
                    "Campaigns"
                ],
                "summary": "Удаление рекламной кампании",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/time/advance": {
            "post": {
                "description": "Устанавливает текущий день в системе в заданную дату",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Time"
                ],
                "summary": "Установка текущей даты",
                "parameters": [
                    {
                        "description": "Новый текущий день",
                        "name": "newDate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CurrentDate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CurrentDate"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Campaign": {
            "type": "object",
            "properties": {
                "ad_text": {
                    "type": "string"
                },
                "ad_title": {
                    "type": "string"
                },
                "advertiser_id": {
                    "type": "string"
                },
                "campaign_id": {
                    "type": "string"
                },
                "clicks_limit": {
                    "type": "integer"
                },
                "cost_per_click": {
                    "type": "number"
                },
                "cost_per_impression": {
                    "type": "number"
                },
                "end_date": {
                    "type": "integer"
                },
                "impressions_limit": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "integer"
                },
                "targeting": {
                    "$ref": "#/definitions/domain.Targeting"
                }
            }
        },
        "domain.CampaignRequest": {
            "type": "object",
            "properties": {
                "ad_text": {
                    "type": "string"
                },
                "ad_title": {
                    "type": "string"
                },
                "clicks_limit": {
                    "type": "integer"
                },
                "cost_per_click": {
                    "type": "number"
                },
                "cost_per_impression": {
                    "type": "number"
                },
                "end_date": {
                    "type": "integer"
                },
                "impressions_limit": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "integer"
                },
                "targeting": {
                    "$ref": "#/definitions/domain.Targeting"
                }
            }
        },
...
        "domain.CurrentDate": {
            "type": "object",
            "properties": {
                "current_date": {
                    "type": "integer"
                }
            }
        },
        "domain.GenerateAdTextResponse": {
            "type": "object",
            "properties": {
                "ad_text": {
                    "type": "string"
                }
            }
        },
...
        "domain.SwitchModerationResponse": {
            "type": "object",
            "properties": {
                "is_moderated": {
                    "type": "boolean"
                }
            }
        },
        "domain.Targeting": {
            "type": "object",
            "properties": {
                "age_from": {
                    "type": "integer"
                },
                "age_to": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                }
            }
        },
...
    }
}